opencv/interview_process/interview_process_project/mycareerinterviewprocess/requirement.txt
asgiref==3.7.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
colorama==0.4.6
decorator==4.4.2
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dlib==19.24.2
dnspython==2.4.2
drf-yasg==1.21.7
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.0
google-ai-generativelanguage==0.3.3
google-api-core==2.14.0
google-auth==2.24.0
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gTTS==2.4.0
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
language-tool-python==2.7.1
MarkupSafe==2.1.3
md5checker==0.2.1
md5hash==0.0.6
mongoengine==0.27.0
moviepy==1.0.3
nltk==3.8.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pay-ccavenue==0.1.6
Pillow==10.1.0
proglog==0.1.10
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.12.0
pymongo==3.12.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
SpeechRecognition==3.10.0
sqlparse==0.2.4
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
